+++ android/src/main/kotlin/co/quis/flutter_contacts/properties/Phone.kt
@@ -2,6 +2,7 @@ package co.quis.flutter_contacts.properties
 
 data class Phone(
     var number: String,
+    var normalizedNumber: String,
     // one of: assistant, callback, car, companyMain, faxHome, faxOther, faxWork, home,
     // iPhone, isdn, main, mms, mobile, pager, radio, school, telex, ttyTtd, work,
     // workMobile, workPager, other, custom
@@ -12,6 +13,7 @@ data class Phone(
     companion object {
         fun fromMap(m: Map<String, Any>): Phone = Phone(
             m["number"] as String,
+            m["normalizedNumber"] as String,
             m["label"] as String,
             m["customLabel"] as String,
             m["isPrimary"] as Boolean
@@ -20,6 +22,7 @@ data class Phone(
 
     fun toMap(): Map<String, Any> = mapOf(
         "number" to number,
+        "normalizedNumber" to normalizedNumber,
         "label" to label,
         "customLabel" to customLabel,
         "isPrimary" to isPrimary